requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
jishaku==2.5.0
